Feat categoery

+++ server/src/Utils/seed.ts
@@ -1,19 +1,49 @@
 import ProductService from "../services/ProductService";
-import products from "../seedData/products.sample.json";
 import { IProduct } from "../model/Product.Model";
 import mongoose from "mongoose";
+import { rand, randBoolean, randEmail, randFood, randFullName, randImg, randNumber, randParagraph, randSentence } from '@ngneat/falso';
+import envconf from "../envconf";
+
+
+let categories = [
+  "Fruits",
+  "Vegetables",
+  "Meat",
+  "Dairy",
+  "Bakery",
+  "Beverages",
+  "Bread",
+  "Canned Goods",
+  "Condiments",
+  "Cereals",
+  "Desserts",
+]
+
 
 
 async function fillProducts() {
   const allProducts = await ProductService.getAllProducts();
 
   if (allProducts.length === 0) {
-    let index = 0
-    for (const product of products) {
-      await ProductService.createProduct(product as IProduct);
-      index++;
-      if (index % 200 === 0)
-        console.log(`${index} products added`);
+    
+    for (const i of Array(1000).fill(1).map((x, i) => i+1)) {
+      const randProduct :IProduct= {
+        name: randFood(),
+        rate: randNumber({ min: 1, max: 5 }),
+        price: randNumber({ min: 10, max: 1000 }),
+        shortDescription: randSentence(),
+        availability: randBoolean(),
+        imgURL: `http://localhost:3000/img/random.png?seed=${randFood().replaceAll(' ', '+')}`,
+        weight: randNumber({ min: 1, max: 1000 }),
+        availableInventory: randNumber({ min: 0, max: 100 }),
+        longDescription: randParagraph(),
+        productInformation: randParagraph(),
+        categoryName: rand(categories),
+      }
+      
+      await ProductService.createProduct(randProduct);
+      if (i % 200 === 0)
+        console.log(`${i} products added`);
     }
   }
 
@@ -39,14 +69,16 @@ async function resetAll() {
     return;
   }
   else {
-    await mongoose.connection.dropDatabase();
-    console.log("Database dropped");
+    if (envconf.databaseReset){
+      await mongoose.connection.dropDatabase();
+      console.log("Database dropped");
+    }
   }
 }
 
 
 async function main() {
-  // await resetAll();
+  await resetAll();
   await fillAll();
 }
